requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-nested-admin==4.0.2
django-phonenumber-field==7.3.0
django-resized==1.0.2
django-rest-framework==0.1.0
django-rest-registration==0.8.3
django-super-inlines==0.2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
google-auth==2.28.2
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
